[test] create 'test' sub-package
- moved test_compression_integrity and test_call_depth to test.util module,
  and rename 'test_*' to 'check_*' to avoid confusions with pytest test cases

- move testDummy to test.dummy
